+++ packages/viking-ui/.storybook/main.ts
@@ -1,17 +1,13 @@
-import type { StorybookConfig } from "@storybook/html";
+import type { StorybookConfig } from "@storybook/html-vite";
 
 const config: StorybookConfig = {
   stories: ["../.storybook/stories/**/*.stories.ts"],
-  addons: [
-    "@storybook/addon-essentials",
-    "@storybook/addon-a11y",
-    "@chromatic-com/storybook",
-  ],
+  addons: ["@storybook/addon-essentials", "@storybook/addon-a11y"],
   core: {
     disableTelemetry: true,
   },
   framework: {
-    name: "@storybook/html",
+    name: "@storybook/html-vite",
     options: {},
   },
   staticDirs: ["../dist"],